Refactor per review
