+++ tests/extra_python_package/test_files.py
@@ -58,6 +58,7 @@ detail_headers = {
     "include/pybind11/detail/type_caster_base.h",
     "include/pybind11/detail/typeid.h",
     "include/pybind11/detail/value_and_holder.h",
+    "include/pybind11/detail/exception_translation.h",
 }
 
 eigen_headers = {
